+++ src/main/java/co/plageopa/repository/UserRepository.java
@@ -5,8 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.plageopa.domain.Usuario;
-
-public interface UserRepository extends JpaRepository<Usuario, Long> {
+public interface UserRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByNombre(String nombre);
     Optional<Usuario> findByCorreo(String correo);
     boolean existsByNombre(String nombre);
